Working

+++ controllers/Admin/helpCenterController.js
@@ -1,22 +1,26 @@
 const helpCenterModel = require('../../models/helpCenter')
-const AppError = require("../Admin/errorController");
+const AppError = require('../Admin/errorController')
 
 exports.createHelpCenter = async (req, res, next) => {
   try {
-    const id = req.user._id
-    const { description, issue, others } = req.body
-    if (!description) {
-      throw new AppError(400, 'All the fields are required')
+    if (req.perm.helpCenter === 'write') {
+      const id = req.user._id
+      const { description, issue, others } = req.body
+      if (!description) {
+        throw new AppError(400, 'All the fields are required')
+      } else {
+        await helpCenterModel.create({
+          userId: id,
+          description: description,
+          issue: issue,
+          others: others
+        })
+        res
+          .status(201)
+          .json({ success: true, message: 'help center created successful' })
+      }
     } else {
-      await helpCenterModel.create({
-        userId: id,
-        description: description,
-        issue: issue,
-        others: others
-      })
-      res
-        .status(201)
-        .json({ success: true, message: 'help center created successful' })
+      throw new AppError(400, 'You are not authorized')
     }
   } catch (err) {
     next(err)
@@ -24,41 +28,58 @@ exports.createHelpCenter = async (req, res, next) => {
 }
 exports.getAllHelpCenter = async (req, res, next) => {
   try {
+    if (req.perm.helpCenter === 'write' || req.perm.helpCenter === 'read') {
+      let status = 'in-review'
+      if (req.body.status) {
+        status = req.body.status
+      }
+      //  const {status}=req.body.status
+      var page = 1
+      if (req.query.page) {
+        page = req.query.page
+      }
+      var limit = 12
+      const allList = await helpCenterModel.find({ status: status }).count()
+      var num = allList / limit
+      var fixedNum = num.toFixed()
+      var totalPage = fixedNum
+      if (num > fixedNum) {
+        totalPage++
+      }
+      const result = await helpCenterModel
+        .find({ status: status })
+        .populate('userId')
+        .limit(limit * 1)
+        .skip((page - 1) * limit)
+        .exec()
 
-    let status="in-review"
-    if(req.body.status){
-         status=req.body.status
+      res
+        .status(201)
+        .json({
+          success: true,
+          message: 'list of all help data',
+          data: result,
+          totalPage: totalPage
+        })
+    } else {
+      throw new AppError(400, 'You are not authorized')
     }
-  //  const {status}=req.body.status
-    var page = 1
-    if (req.query.page) {
-      page = req.query.page
-    }
-    var limit = 12
-    const allList = await helpCenterModel.find({"status":status}).count()
-    var num = allList / limit
-    var fixedNum = num.toFixed()
-    var totalPage = fixedNum
-    if (num > fixedNum) {
-      totalPage++
-    }
-    const result = await helpCenterModel.find({"status":status}).populate("userId") .limit(limit * 1)
-    .skip((page - 1) * limit)
-    .exec()
-
-    res
-      .status(201)
-      .json({ success: true, message: 'list of all help data', data: result,totalPage:totalPage })
   } catch (err) {
     next(err)
   }
 }
 exports.deleteHelpCenter = async (req, res, next) => {
   try {
-    const id = req.params.id
+    if (req.perm.helpCenter === 'write') {
+      const id = req.params.id
 
-    await helpCenterModel.findByIdAndDelete({ _id: id })
-    res.status(201).json({ success: true, message: 'data deleted successful' })
+      await helpCenterModel.findByIdAndDelete({ _id: id })
+      res
+        .status(201)
+        .json({ success: true, message: 'data deleted successful' })
+    } else {
+      throw new AppError(400, 'You are not authorized')
+    }
   } catch (err) {
     next(err)
   }
@@ -66,18 +87,22 @@ exports.deleteHelpCenter = async (req, res, next) => {
 
 exports.updateHelpCenter = async (req, res, next) => {
   try {
-    const id = req.params.id
-    const { resolution } = req.body
-    if (!resolution) {
-      throw new AppError(400, 'Please provide resolution')
+    if (req.perm.helpCenter != 'write' || req.perm.helpCenter != 'read') {
+      const id = req.params.id
+      const { resolution } = req.body
+      if (!resolution) {
+        throw new AppError(400, 'Please provide resolution')
+      } else {
+        var result = await helpCenterModel.findOne({ _id: id })
+        result.resolution = resolution
+        result.status = 'solved'
+        await result.save()
+        res
+          .status(201)
+          .json({ success: true, message: 'data updated successful' })
+      }
     } else {
-      var result = await helpCenterModel.findOne({ _id: id })
-      result.resolution = resolution
-      result.status = 'solved'
-      await result.save()
-      res
-        .status(201)
-        .json({ success: true, message: 'data updated successful' })
+      throw new AppError(400, 'You are not authorized')
     }
   } catch (err) {
     next(err)
@@ -86,21 +111,22 @@ exports.updateHelpCenter = async (req, res, next) => {
 
 exports.getUserHelpCenter = async (req, res, next) => {
   try {
-    const id = req.user._id
-    const result = await helpCenterModel.find({ "userId": id })
-    if (result.length == 0) {
-      throw new AppError(400, 'Data not found')
-    } else {
-      res
-        .status(201)
-        .json({
+    if (req.perm.helpCenter === 'write' || req.perm.helpCenter === 'read') {
+      const id = req.user._id
+      const result = await helpCenterModel.find({ userId: id })
+      if (result.length == 0) {
+        throw new AppError(400, 'Data not found')
+      } else {
+        res.status(201).json({
           success: true,
           message: 'data updated successful',
           data: result
         })
+      }
+    } else {
+      throw new AppError(400, 'You are not authorized')
     }
   } catch (err) {
     next(err)
   }
 }
-
